requirements.txt
Flask==2.3.3
requests==2.31.0
APScheduler==3.8.0
MarkupSafe==2.1.3
python-dotenv==1.0.0
cachetools==5.3.1